Allow decimal values for Carbs, Fats and Protein entries integers -> Double. Localized entries with , gets converted to . for calculations maxCarbs & maxFatProtein

+++ LoopFollow/ViewControllers/RemoteSettingsViewController.swift
@@ -306,7 +306,8 @@ class RemoteSettingsViewController: FormViewController {
             }
         }.onChange { [weak self] row in
             guard let value = row.value else { return }
-            UserDefaultsRepository.maxCarbs.value = Int(value)
+            //UserDefaultsRepository.maxCarbs.value = Int(value)
+            UserDefaultsRepository.maxCarbs.value = Double(value)
         }
         
         <<< StepperRow("maxFatProtein") { row in
@@ -321,7 +322,8 @@ class RemoteSettingsViewController: FormViewController {
             }
         }.onChange { [weak self] row in
             guard let value = row.value else { return }
-            UserDefaultsRepository.maxFatProtein.value = Int(value)
+            //UserDefaultsRepository.maxFatProtein.value = Int(value)
+            UserDefaultsRepository.maxFatProtein.value = Double(value)
         }
         
         <<< StepperRow("maxBolus") { row in
